+++ mocks/service_container.go
@@ -34,7 +34,7 @@ type MockServiceContainer struct {
 func NewMockServiceContainer() *MockServiceContainer {
 	return &MockServiceContainer{
 		GetFunc: &ServiceContainerGetFunc{
-			defaultHook: func(string) (interface{}, error) {
+			defaultHook: func(interface{}) (interface{}, error) {
 				return nil, nil
 			},
 		},
@@ -44,17 +44,17 @@ func NewMockServiceContainer() *MockServiceContainer {
 			},
 		},
 		MustGetFunc: &ServiceContainerMustGetFunc{
-			defaultHook: func(string) interface{} {
+			defaultHook: func(interface{}) interface{} {
 				return nil
 			},
 		},
 		MustSetFunc: &ServiceContainerMustSetFunc{
-			defaultHook: func(string, interface{}) {
+			defaultHook: func(interface{}, interface{}) {
 				return
 			},
 		},
 		SetFunc: &ServiceContainerSetFunc{
-			defaultHook: func(string, interface{}) error {
+			defaultHook: func(interface{}, interface{}) error {
 				return nil
 			},
 		},
@@ -87,15 +87,15 @@ func NewMockServiceContainerFrom(i service.ServiceContainer) *MockServiceContain
 // ServiceContainerGetFunc describes the behavior when the Get method of the
 // parent MockServiceContainer instance is invoked.
 type ServiceContainerGetFunc struct {
-	defaultHook func(string) (interface{}, error)
-	hooks       []func(string) (interface{}, error)
+	defaultHook func(interface{}) (interface{}, error)
+	hooks       []func(interface{}) (interface{}, error)
 	history     []ServiceContainerGetFuncCall
 	mutex       sync.Mutex
 }
 
 // Get delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockServiceContainer) Get(v0 string) (interface{}, error) {
+func (m *MockServiceContainer) Get(v0 interface{}) (interface{}, error) {
 	r0, r1 := m.GetFunc.nextHook()(v0)
 	m.GetFunc.appendCall(ServiceContainerGetFuncCall{v0, r0, r1})
 	return r0, r1
@@ -104,7 +104,7 @@ func (m *MockServiceContainer) Get(v0 string) (interface{}, error) {
 // SetDefaultHook sets function that is called when the Get method of the
 // parent MockServiceContainer instance is invoked and the hook queue is
 // empty.
-func (f *ServiceContainerGetFunc) SetDefaultHook(hook func(string) (interface{}, error)) {
+func (f *ServiceContainerGetFunc) SetDefaultHook(hook func(interface{}) (interface{}, error)) {
 	f.defaultHook = hook
 }
 
@@ -112,7 +112,7 @@ func (f *ServiceContainerGetFunc) SetDefaultHook(hook func(string) (interface{},
 // Get method of the parent MockServiceContainer instance invokes the hook
 // at the front of the queue and discards it. After the queue is empty, the
 // default hook function is invoked for any future action.
-func (f *ServiceContainerGetFunc) PushHook(hook func(string) (interface{}, error)) {
+func (f *ServiceContainerGetFunc) PushHook(hook func(interface{}) (interface{}, error)) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -121,7 +121,7 @@ func (f *ServiceContainerGetFunc) PushHook(hook func(string) (interface{}, error
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *ServiceContainerGetFunc) SetDefaultReturn(r0 interface{}, r1 error) {
-	f.SetDefaultHook(func(string) (interface{}, error) {
+	f.SetDefaultHook(func(interface{}) (interface{}, error) {
 		return r0, r1
 	})
 }
@@ -129,12 +129,12 @@ func (f *ServiceContainerGetFunc) SetDefaultReturn(r0 interface{}, r1 error) {
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *ServiceContainerGetFunc) PushReturn(r0 interface{}, r1 error) {
-	f.PushHook(func(string) (interface{}, error) {
+	f.PushHook(func(interface{}) (interface{}, error) {
 		return r0, r1
 	})
 }
 
-func (f *ServiceContainerGetFunc) nextHook() func(string) (interface{}, error) {
+func (f *ServiceContainerGetFunc) nextHook() func(interface{}) (interface{}, error) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -169,7 +169,7 @@ func (f *ServiceContainerGetFunc) History() []ServiceContainerGetFuncCall {
 type ServiceContainerGetFuncCall struct {
 	// Arg0 is the value of the 1st argument passed to this method
 	// invocation.
-	Arg0 string
+	Arg0 interface{}
 	// Result0 is the value of the 1st result returned from this method
 	// invocation.
 	Result0 interface{}
@@ -296,15 +296,15 @@ func (c ServiceContainerInjectFuncCall) Results() []interface{} {
 // ServiceContainerMustGetFunc describes the behavior when the MustGet
 // method of the parent MockServiceContainer instance is invoked.
 type ServiceContainerMustGetFunc struct {
-	defaultHook func(string) interface{}
-	hooks       []func(string) interface{}
+	defaultHook func(interface{}) interface{}
+	hooks       []func(interface{}) interface{}
 	history     []ServiceContainerMustGetFuncCall
 	mutex       sync.Mutex
 }
 
 // MustGet delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockServiceContainer) MustGet(v0 string) interface{} {
+func (m *MockServiceContainer) MustGet(v0 interface{}) interface{} {
 	r0 := m.MustGetFunc.nextHook()(v0)
 	m.MustGetFunc.appendCall(ServiceContainerMustGetFuncCall{v0, r0})
 	return r0
@@ -313,7 +313,7 @@ func (m *MockServiceContainer) MustGet(v0 string) interface{} {
 // SetDefaultHook sets function that is called when the MustGet method of
 // the parent MockServiceContainer instance is invoked and the hook queue is
 // empty.
-func (f *ServiceContainerMustGetFunc) SetDefaultHook(hook func(string) interface{}) {
+func (f *ServiceContainerMustGetFunc) SetDefaultHook(hook func(interface{}) interface{}) {
 	f.defaultHook = hook
 }
 
@@ -321,7 +321,7 @@ func (f *ServiceContainerMustGetFunc) SetDefaultHook(hook func(string) interface
 // MustGet method of the parent MockServiceContainer instance invokes the
 // hook at the front of the queue and discards it. After the queue is empty,
 // the default hook function is invoked for any future action.
-func (f *ServiceContainerMustGetFunc) PushHook(hook func(string) interface{}) {
+func (f *ServiceContainerMustGetFunc) PushHook(hook func(interface{}) interface{}) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -330,7 +330,7 @@ func (f *ServiceContainerMustGetFunc) PushHook(hook func(string) interface{}) {
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *ServiceContainerMustGetFunc) SetDefaultReturn(r0 interface{}) {
-	f.SetDefaultHook(func(string) interface{} {
+	f.SetDefaultHook(func(interface{}) interface{} {
 		return r0
 	})
 }
@@ -338,12 +338,12 @@ func (f *ServiceContainerMustGetFunc) SetDefaultReturn(r0 interface{}) {
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *ServiceContainerMustGetFunc) PushReturn(r0 interface{}) {
-	f.PushHook(func(string) interface{} {
+	f.PushHook(func(interface{}) interface{} {
 		return r0
 	})
 }
 
-func (f *ServiceContainerMustGetFunc) nextHook() func(string) interface{} {
+func (f *ServiceContainerMustGetFunc) nextHook() func(interface{}) interface{} {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -378,7 +378,7 @@ func (f *ServiceContainerMustGetFunc) History() []ServiceContainerMustGetFuncCal
 type ServiceContainerMustGetFuncCall struct {
 	// Arg0 is the value of the 1st argument passed to this method
 	// invocation.
-	Arg0 string
+	Arg0 interface{}
 	// Result0 is the value of the 1st result returned from this method
 	// invocation.
 	Result0 interface{}
@@ -399,15 +399,15 @@ func (c ServiceContainerMustGetFuncCall) Results() []interface{} {
 // ServiceContainerMustSetFunc describes the behavior when the MustSet
 // method of the parent MockServiceContainer instance is invoked.
 type ServiceContainerMustSetFunc struct {
-	defaultHook func(string, interface{})
-	hooks       []func(string, interface{})
+	defaultHook func(interface{}, interface{})
+	hooks       []func(interface{}, interface{})
 	history     []ServiceContainerMustSetFuncCall
 	mutex       sync.Mutex
 }
 
 // MustSet delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockServiceContainer) MustSet(v0 string, v1 interface{}) {
+func (m *MockServiceContainer) MustSet(v0 interface{}, v1 interface{}) {
 	m.MustSetFunc.nextHook()(v0, v1)
 	m.MustSetFunc.appendCall(ServiceContainerMustSetFuncCall{v0, v1})
 	return
@@ -416,7 +416,7 @@ func (m *MockServiceContainer) MustSet(v0 string, v1 interface{}) {
 // SetDefaultHook sets function that is called when the MustSet method of
 // the parent MockServiceContainer instance is invoked and the hook queue is
 // empty.
-func (f *ServiceContainerMustSetFunc) SetDefaultHook(hook func(string, interface{})) {
+func (f *ServiceContainerMustSetFunc) SetDefaultHook(hook func(interface{}, interface{})) {
 	f.defaultHook = hook
 }
 
@@ -424,7 +424,7 @@ func (f *ServiceContainerMustSetFunc) SetDefaultHook(hook func(string, interface
 // MustSet method of the parent MockServiceContainer instance invokes the
 // hook at the front of the queue and discards it. After the queue is empty,
 // the default hook function is invoked for any future action.
-func (f *ServiceContainerMustSetFunc) PushHook(hook func(string, interface{})) {
+func (f *ServiceContainerMustSetFunc) PushHook(hook func(interface{}, interface{})) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -433,7 +433,7 @@ func (f *ServiceContainerMustSetFunc) PushHook(hook func(string, interface{})) {
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *ServiceContainerMustSetFunc) SetDefaultReturn() {
-	f.SetDefaultHook(func(string, interface{}) {
+	f.SetDefaultHook(func(interface{}, interface{}) {
 		return
 	})
 }
@@ -441,12 +441,12 @@ func (f *ServiceContainerMustSetFunc) SetDefaultReturn() {
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *ServiceContainerMustSetFunc) PushReturn() {
-	f.PushHook(func(string, interface{}) {
+	f.PushHook(func(interface{}, interface{}) {
 		return
 	})
 }
 
-func (f *ServiceContainerMustSetFunc) nextHook() func(string, interface{}) {
+func (f *ServiceContainerMustSetFunc) nextHook() func(interface{}, interface{}) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -481,7 +481,7 @@ func (f *ServiceContainerMustSetFunc) History() []ServiceContainerMustSetFuncCal
 type ServiceContainerMustSetFuncCall struct {
 	// Arg0 is the value of the 1st argument passed to this method
 	// invocation.
-	Arg0 string
+	Arg0 interface{}
 	// Arg1 is the value of the 2nd argument passed to this method
 	// invocation.
 	Arg1 interface{}
@@ -502,15 +502,15 @@ func (c ServiceContainerMustSetFuncCall) Results() []interface{} {
 // ServiceContainerSetFunc describes the behavior when the Set method of the
 // parent MockServiceContainer instance is invoked.
 type ServiceContainerSetFunc struct {
-	defaultHook func(string, interface{}) error
-	hooks       []func(string, interface{}) error
+	defaultHook func(interface{}, interface{}) error
+	hooks       []func(interface{}, interface{}) error
 	history     []ServiceContainerSetFuncCall
 	mutex       sync.Mutex
 }
 
 // Set delegates to the next hook function in the queue and stores the
 // parameter and result values of this invocation.
-func (m *MockServiceContainer) Set(v0 string, v1 interface{}) error {
+func (m *MockServiceContainer) Set(v0 interface{}, v1 interface{}) error {
 	r0 := m.SetFunc.nextHook()(v0, v1)
 	m.SetFunc.appendCall(ServiceContainerSetFuncCall{v0, v1, r0})
 	return r0
@@ -519,7 +519,7 @@ func (m *MockServiceContainer) Set(v0 string, v1 interface{}) error {
 // SetDefaultHook sets function that is called when the Set method of the
 // parent MockServiceContainer instance is invoked and the hook queue is
 // empty.
-func (f *ServiceContainerSetFunc) SetDefaultHook(hook func(string, interface{}) error) {
+func (f *ServiceContainerSetFunc) SetDefaultHook(hook func(interface{}, interface{}) error) {
 	f.defaultHook = hook
 }
 
@@ -527,7 +527,7 @@ func (f *ServiceContainerSetFunc) SetDefaultHook(hook func(string, interface{}) 
 // Set method of the parent MockServiceContainer instance invokes the hook
 // at the front of the queue and discards it. After the queue is empty, the
 // default hook function is invoked for any future action.
-func (f *ServiceContainerSetFunc) PushHook(hook func(string, interface{}) error) {
+func (f *ServiceContainerSetFunc) PushHook(hook func(interface{}, interface{}) error) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -536,7 +536,7 @@ func (f *ServiceContainerSetFunc) PushHook(hook func(string, interface{}) error)
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *ServiceContainerSetFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(string, interface{}) error {
+	f.SetDefaultHook(func(interface{}, interface{}) error {
 		return r0
 	})
 }
@@ -544,12 +544,12 @@ func (f *ServiceContainerSetFunc) SetDefaultReturn(r0 error) {
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *ServiceContainerSetFunc) PushReturn(r0 error) {
-	f.PushHook(func(string, interface{}) error {
+	f.PushHook(func(interface{}, interface{}) error {
 		return r0
 	})
 }
 
-func (f *ServiceContainerSetFunc) nextHook() func(string, interface{}) error {
+func (f *ServiceContainerSetFunc) nextHook() func(interface{}, interface{}) error {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -584,7 +584,7 @@ func (f *ServiceContainerSetFunc) History() []ServiceContainerSetFuncCall {
 type ServiceContainerSetFuncCall struct {
 	// Arg0 is the value of the 1st argument passed to this method
 	// invocation.
-	Arg0 string
+	Arg0 interface{}
 	// Arg1 is the value of the 2nd argument passed to this method
 	// invocation.
 	Arg1 interface{}